+++ bugsnag-android-core/src/main/java/com/bugsnag/android/ImmutableConfig.kt
@@ -3,6 +3,7 @@ package com.bugsnag.android
 import android.content.Context
 import android.content.pm.ApplicationInfo
 import android.content.pm.PackageManager
+import java.io.File
 
 internal data class ImmutableConfig(
     val apiKey: String,
@@ -26,7 +27,8 @@ internal data class ImmutableConfig(
     val logger: Logger,
     val maxBreadcrumbs: Int,
     val maxPersistedEvents: Int,
-    val maxPersistedSessions: Int
+    val maxPersistedSessions: Int,
+    val persistenceDirectory: File
 ) {
 
     /**
@@ -82,7 +84,8 @@ internal fun convertToImmutableConfig(
         maxBreadcrumbs = config.maxBreadcrumbs,
         maxPersistedEvents = config.maxPersistedEvents,
         maxPersistedSessions = config.maxPersistedSessions,
-        enabledBreadcrumbTypes = config.enabledBreadcrumbTypes?.toSet()
+        enabledBreadcrumbTypes = config.enabledBreadcrumbTypes?.toSet(),
+        persistenceDirectory = config.persistenceDirectory!!
     )
 }
 
@@ -131,6 +134,10 @@ internal fun sanitiseConfiguration(
     @Suppress("SENSELESS_COMPARISON")
     if (configuration.delivery == null) {
         configuration.delivery = DefaultDelivery(connectivity, configuration.logger!!)
+    }
+
+    if (configuration.persistenceDirectory == null) {
+        configuration.persistenceDirectory = appContext.cacheDir
     }
     return convertToImmutableConfig(configuration, buildUuid)
 }