+++ bugsnag-android-core/src/test/java/com/bugsnag/android/EventFilenameTest.kt
@@ -10,7 +10,6 @@ import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.mockito.Mock
-import org.mockito.Mockito
 import org.mockito.Mockito.*
 import org.mockito.junit.MockitoJUnitRunner
 import java.io.File
@@ -43,7 +42,6 @@ class EventFilenameTest {
     fun setUp() {
         eventStore = EventStore(
             config,
-            context,
             NoopLogger,
             Notifier(),
             null
@@ -100,8 +98,8 @@ class EventFilenameTest {
 
     @Test
     fun regularJvmEventName() {
-        val filename = eventStore.getFilename(event, "my-uuid-123", null, 1504255147933, "/errors/")
-        assertEquals("/errors/1504255147933_0000111122223333aaaabbbbcccc9999_my-uuid-123.json", filename)
+        val filename = eventStore.getFilename(event, "my-uuid-123", null, 1504255147933)
+        assertEquals("1504255147933_0000111122223333aaaabbbbcccc9999_my-uuid-123.json", filename)
     }
 
     /**
@@ -110,8 +108,8 @@ class EventFilenameTest {
     @Test
     fun startupCrashJvmEventName() {
         `when`(app.duration).thenReturn(1000)
-        val filename = eventStore.getFilename(event, "my-uuid-123", null, 1504255147933, "/errors/")
-        assertEquals("/errors/1504255147933_0000111122223333aaaabbbbcccc9999_my-uuid-123_startupcrash.json", filename)
+        val filename = eventStore.getFilename(event, "my-uuid-123", null, 1504255147933)
+        assertEquals("1504255147933_0000111122223333aaaabbbbcccc9999_my-uuid-123_startupcrash.json", filename)
     }
 
     /**
@@ -120,21 +118,21 @@ class EventFilenameTest {
     @Test
     fun nonStartupCrashCrashJvmEventName() {
         `when`(app.duration).thenReturn(10000)
-        val filename = eventStore.getFilename(event, "my-uuid-123", null, 1504255147933, "/errors/")
-        assertEquals("/errors/1504255147933_0000111122223333aaaabbbbcccc9999_my-uuid-123.json", filename)
+        val filename = eventStore.getFilename(event, "my-uuid-123", null, 1504255147933)
+        assertEquals("1504255147933_0000111122223333aaaabbbbcccc9999_my-uuid-123.json", filename)
     }
 
     @Test
     fun ndkEventName() {
         val filename = eventStore.getFilename("{}", "my-uuid-123",
-            "0000111122223333aaaabbbbcccc9999", 1504255147933, "/errors/")
-        assertEquals("/errors/1504255147933_0000111122223333aaaabbbbcccc9999_my-uuid-123not-jvm.json", filename)
+            "0000111122223333aaaabbbbcccc9999", 1504255147933)
+        assertEquals("1504255147933_0000111122223333aaaabbbbcccc9999_my-uuid-123not-jvm.json", filename)
     }
 
     @Test
     fun ndkEventNameNoApiKey() {
-        val filename = eventStore.getFilename("{}", "my-uuid-123", "", 1504255147933, "/errors/")
-        assertEquals("/errors/1504255147933_5d1ec5bd39a74caa1267142706a7fb21_my-uuid-123not-jvm.json", filename)
+        val filename = eventStore.getFilename("{}", "my-uuid-123", "", 1504255147933)
+        assertEquals("1504255147933_5d1ec5bd39a74caa1267142706a7fb21_my-uuid-123not-jvm.json", filename)
     }
 
     @Test